% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_separate.R
\name{coarse_gather}
\alias{coarse_gather}
\title{Gather more coarse mesh}
\usage{
coarse_gather(meshcode, distinct = FALSE)
}
\arguments{
\item{meshcode}{\code{character}. mesh code}

\item{distinct}{return unique meshcodes}
}
\value{
character vector

meshcode as \code{character}
}
\description{
Return coarse gather mesh codes
}
\examples{
m <- c("493214294", "493214392", "493215203", "493215301")
coarse_gather(m)
coarse_gather(coarse_gather(m))
coarse_gather(coarse_gather(m), distinct = TRUE)
}
