% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_convert.R
\name{mesh_convert}
\alias{mesh_convert}
\title{Mesh unit converter}
\usage{
mesh_convert(meshcode = NULL, to_mesh_size = NULL)
}
\arguments{
\item{meshcode}{\code{character}. mesh code}

\item{to_mesh_size}{target mesh type. From 80km to 0.125km. If \code{NULL},
the meshcode of one small scale will be returned. If it is the same as the original size,
the meshcode of the input will be returne.}
}
\description{
Return different meshcode values included in the mesh.
}
