% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_mesh.R
\name{coords_to_mesh}
\alias{coords_to_mesh}
\title{Convert from coordinate to mesh code}
\usage{
coords_to_mesh(longitude, latitude, mesh_size = "1km", ...)
}
\arguments{
\item{longitude}{longitude that approximately to .120.0 to 154.0 (\code{double})}

\item{latitude}{latitude that approximately to 20.0 to 46.0 (\code{double})}

\item{mesh_size}{mesh type. From 80km to 125m}

\item{...}{other parameters}
}
\value{
mesh code (default 3rd meshcode aka 1km mesh)
}
\description{
From coordinate to mesh codes.
}
\examples{
coords_to_mesh(141.3468, 43.06462, mesh_size = "10km")
coords_to_mesh(139.6917, 35.68949, mesh_size = "250m")
coords_to_mesh(139.71475, 35.70078)
}
\references{
Akio Takenaka: \url{http://takenaka-akio.org/etc/j_map/index.html}
}
\seealso{
\code{\link[=mesh_to_coords]{mesh_to_coords()}} for convert from meshcode to coordinates
}
