% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_separate.R
\name{detect_mesh}
\alias{detect_mesh}
\title{Detect file scale mesh code}
\usage{
detect_mesh(meshcode, lat, long)
}
\arguments{
\item{meshcode}{mesh code}

\item{lat}{latitude}

\item{long}{longitude}
}

