% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mesh.R
\name{export_mesh}
\alias{export_mesh}
\title{Export meshcode to geometry}
\usage{
export_mesh(meshcode)
}
\arguments{
\item{meshcode}{mesh code}

\item{...}{other parameters}
}
\value{
sf object
}
\description{
Convert and export meshcode area to sf polygon.
}
\examples{
export_mesh(6441427712)
}
