% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_separate.R
\name{fine_separate}
\alias{fine_separate}
\title{Separate more fine mesh order}
\usage{
fine_separate(mesh = NULL, ...)
}
\arguments{
\item{mesh}{mesh code}

\item{...}{other parameters for paste}
}
\value{
character vector
}
\description{
Return contains fine mesh codes
}
\examples{
fine_separate(52350400)
fine_separate(52350400)
fine_separate(52350400)
}
