% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mesh.R
\name{is_mesh}
\alias{is_mesh}
\alias{is_meshcode}
\alias{is_corner}
\title{Predict meshcode format and positions}
\usage{
is_meshcode(meshcode)

is_corner(meshcode)
}
\arguments{
\item{meshcode}{\code{integer}. mesh code}
}
\description{
Predict meshcode format and positions for utility and certain.
}
