% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpmesh.pref}
\alias{jpmesh.pref}
\title{Prefecture mesh code include lat-lon}
\format{A data frame with 2307485 rows and 8 variables:
\itemize{
  \item{jiscode: }
  \item{long: longitude}
  \item{lat: latitude}
  \item{order}
  \item{group}
  \item{id}
  \item{city_code}
  \item{city_name}
}}
\usage{
jpmesh.pref
}
\keyword{datasets}

