% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpmesh.prefs.tidy}
\alias{jpmesh.prefs.tidy}
\title{}
\format{A data frame with 2010060 rows and 6 variables:
\itemize{
  \item{jiscode: }
  \item{long: longitude}
  \item{lat: latitude}
  \item{order}
  \item{group}
  \item{id}
}}
\usage{
jpmesh.prefs.tidy
}
\keyword{datasets}

