% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpnrect}
\alias{jpnrect}
\title{Simple displaed as rectangel for Japan (fortified)}
\format{
A data frame with 235 rows 11 variables:
\itemize{
\item{long}
\item{lat}
\item{order}
\item{hole}
\item{piece}
\item{id}
\item{group}
\item{mesh_code}
\item{latitude}
\item{longitude}
\item{abb_name}
}
}
\usage{
jpnrect
}
\description{
Rectangle Japanese prefectures positions.
}
\examples{
\dontrun{
plot(jpnrect["abb_name"])
}
}
\keyword{datasets}
