% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_mesh.R
\name{latlong_to_meshcode}
\alias{latlong_to_meshcode}
\title{Convert from coordinate to mesh code}
\usage{
latlong_to_meshcode(lat = NULL, long = NULL, order = 3)
}
\arguments{
\item{lat}{numeric. latitude}

\item{long}{numeric. longitude}

\item{order}{integer. mesh type}
}
\value{
mesh code (default 3rd meshcode)
}
\description{
From coordinate to mesh codes.
}
\details{
http://takenaka-akio.org/etc/j_map/index.html
}
\examples{
latlong_to_meshcode(43.06462, 141.3468, order = 3)
latlong_to_meshcode(35.68949, 139.6917, order = 2)
}
\author{
Akio Takenaka
}

