% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_to_mesh.R
\name{latlong_to_sepate_mesh}
\alias{latlong_to_sepate_mesh}
\title{Convert from coordinate to separate mesh code}
\usage{
latlong_to_sepate_mesh(lat = NULL, long = NULL, order = c("harf",
  "quarter"))
}
\arguments{
\item{lat}{numeric. latitude}

\item{order}{choose character harf or quarter}

\item{lon}{numeric. longitude}
}
\value{
separate meshcode
}
\examples{
latlong_to_sepate_mesh(35.442788, 139.301255, order = "harf")
latlong_to_sepate_mesh(35.442893, 139.310654, order = "quarter")
latlong_to_sepate_mesh(35.448767, 139.301706, order = "quarter")
latlong_to_sepate_mesh(35.449011, 139.311340, order = "quarter")
}

