% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_rectangle.R
\name{mesh_area}
\alias{mesh_area}
\title{Rectange mesh grid area}
\usage{
mesh_area(code, order = c("harf", "quarter", "eight"))
}
\arguments{
\item{code}{meshcode}

\item{order}{mesh order}
}
\description{
Rectange mesh grid area
}
\examples{
mesh_area(523504221, order = "harf")
}

