% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_rectangle.R
\name{mesh_rectangle}
\alias{mesh_rectangle}
\title{mesh rectangel check}
\usage{
mesh_rectangle(df, mesh_code = "mesh_code", view = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{mesh_code}{mesh code}

\item{view}{option.}
}
\examples{
\dontrun{
library(dplyr)
mesh_rectangle(pref_mesh(33), mesh_code = "id", view = FALSE)
}
}

