% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mesh_size}
\alias{mesh_size}
\title{Identifer to mesh size}
\usage{
mesh_size(meshcode)
}
\arguments{
\item{meshcode}{\code{character}. mesh code}
}
\description{
Returns a unit object of mesh size for the given number.
}
