% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_latlon.R
\name{mesh_to_coords}
\alias{mesh_to_coords}
\title{Get from mesh code to latitude and longitude}
\usage{
mesh_to_coords(meshcode, ...)
}
\arguments{
\item{meshcode}{numeric. mesh code}

\item{...}{other parameters}
}
\description{
mesh centroid
}
\details{
http://takenaka-akio.org/etc/j_map/index.html
}
\examples{
mesh_to_coords(64414277)
}
\author{
Akio Takenaka
}
