% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_viewer.R
\name{mesh_viewer}
\alias{mesh_viewer}
\title{interactive meshcode check}
\usage{
mesh_viewer(...)
}
\arguments{
\item{...}{other parameters}
}
\description{
Shiny gadgets for jpmesh.
}
\examples{
\dontrun{
mesh_viewer()
}
}
