% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshcode.R
\name{meshcode_vector}
\alias{meshcode_vector}
\alias{meshcode}
\alias{as_meshcode}
\alias{format.meshcode}
\title{Vector of meshcode}
\usage{
meshcode_vector(x = character(), size = double())

meshcode(x)

as_meshcode(x, ...)

\method{format}{meshcode}(x, ...)
}
\arguments{
\item{x}{input meshcode value}

\item{size}{input meshcode size. Default set to \code{NULL}. The decision is
automatically made based on the \code{meshsize}.}

\item{...}{path to another function}
}
\value{
\link{meshcode}
}
\description{
Vector of meshcode
}
\examples{
meshcode("6441")
}
