% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_latlon.R
\name{meshcode_to_latlon}
\alias{meshcode_to_latlon}
\title{Get from mesh code to latitude and longitude}
\usage{
meshcode_to_latlon(code)
}
\arguments{
\item{code}{numeric. mesh code}
}
\description{
mesh centroid
}
\details{
http://takenaka-akio.org/etc/j_map/index.html
}
\examples{
meshcode_to_latlon(64414277)
}
\author{
Akio Takenaka
}

