% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prefecture_mesh}
\alias{prefecture_mesh}
\title{Meshcode include the prefecture}
\format{A data frame with 240 rows 2 variables:
\itemize{
  \item{pref}
  \item{mesh}
}}
\usage{
prefecture_mesh
}
\description{
Japanese prefectures mesh dataset
}
\details{
50km meshcode
}
\keyword{datasets}
