% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmesh.R
\name{rmesh}
\alias{rmesh}
\title{Generate random sample meshcode}
\usage{
rmesh(n, mesh_size = 1)
}
\arguments{
\item{n}{Number of samples}

\item{mesh_size}{Export mesh size from 80km to 1km.}
}
\description{
Generate random sample meshcode
}
\examples{
rmesh(3, mesh_size = 1)
}
