% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_jpmesh}
\alias{sf_jpmesh}
\title{1:200,000 Scale Maps Name with Meshcode of Japan.}
\format{A data frame with 175 rows 9 variables:
\itemize{
\item{meshcode: 80km meshcode}
\item{name: names for map}
\item{name_roman: names for map (roman)}
\item{lng_center: centroid coordiates of mesh}
\item{lat_center: centroid coordiates of mesh}
\item{lng_error: mesh area}
\item{lat_error: mesh area}
\item{type: evalueate value to mesh}
}}
\usage{
sf_jpmesh
}
\description{
Information for the 1:200,000 Scale Maps.
}
\examples{
\dontrun{
plot(sf_jpmesh["name_roman"])
}
}
\keyword{datasets}
