% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administration_mesh.R
\name{administration_mesh}
\alias{administration_mesh}
\title{Extract administration mesh}
\usage{
administration_mesh(code, type = c("city", "prefecture"))
}
\arguments{
\item{code}{administration code}

\item{type}{administration type. Should be one of "prefecture" or "city".}
}
\description{
Extract administration mesh
}
\examples{
\dontrun{
administration_mesh(code = c(35201), type = "city")
administration_mesh(code = c(35), type = "prefecture")
administration_mesh(code = c(33, 34), type = "prefecture")
}
}
