% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_separate.R
\name{detect_mesh}
\alias{detect_mesh}
\title{Detect file scale mesh code}
\usage{
detect_mesh(meshcode, lat, long)
}
\arguments{
\item{meshcode}{mesh code}

\item{lat}{latitude}

\item{long}{longitude}
}
\description{
Find out position.
}
\examples{
detect_mesh(52350422, lat = 34.684176, long = 135.526130)
detect_mesh(523504221, lat = 34.684028, long = 135.529506)
}
