% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mesh.R
\name{export_mesh}
\alias{export_mesh}
\title{Export rectangle mesh to geojson}
\usage{
export_mesh(code)
}
\arguments{
\item{code}{meshcode}
}
\value{
sf
}
\description{
Convert and export meshcode area to geojson string.
}
\examples{
export_mesh(
  code = latlong_to_meshcode(lat = 35.82083, long = 137.0563, order = 3))

library(leaflet)
leaflet() \%>\%
addTiles() \%>\%
setView(lng = 135.0, lat = 38.0, zoom = 3) \%>\% 
addPolygons(data = export_mesh(
code = c(
latlong_to_meshcode(lat = 35.82083, long = 137.0563, order = 3),
latlong_to_meshcode(lat = 38.40052, long = 141.0131, order = 3))))
}
