% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mesh.R
\name{export_meshes}
\alias{export_meshes}
\title{Export meshcode to geometry}
\usage{
export_meshes(meshes)
}
\arguments{
\item{meshes}{mesh code sets}
}
\description{
Convert and export meshcode area to \code{sf}.
}
\examples{
export_meshes("4128")
\dontrun{
find_neighbor_mesh(37250395) \%>\% 
  export_meshes()
}
}
