% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbor_mesh.R
\name{find_neighbor_mesh}
\alias{find_neighbor_mesh}
\title{Find out neighborhood meshes collection}
\usage{
find_neighbor_mesh(mesh = NULL, contains = TRUE)
}
\arguments{
\item{mesh}{meshcode (under the 1km mesh size)}

\item{contains}{logical. contains input meshcode}
}
\description{
Find out neighborhood meshes collection
}
\examples{
find_neighbor_mesh(53394501)
find_neighbor_mesh(533945011)
find_neighbor_mesh(533945011, contains = FALSE)
}
