% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_separate.R
\name{fine_separate}
\alias{fine_separate}
\title{Separate more fine mesh order}
\usage{
fine_separate(meshcode = NULL, .type = "standard", ...)
}
\arguments{
\item{meshcode}{\code{character}. mesh code}

\item{.type}{Specify the \code{subdivision} if you want to get a 100m mesh.}

\item{...}{other parameters for \link[base:paste]{paste}}
}
\value{
\link{meshcode}
}
\description{
Return contains fine mesh codes
}
\examples{
fine_separate("5235")
fine_separate("523504")
fine_separate("52350432")
fine_separate("523504321")
fine_separate("5235043211")
# to 100m mesh code
fine_separate("64414315", .type = "subdivision")
}
