% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_separate.R
\name{fine_separate}
\alias{fine_separate}
\title{Separate more fine mesh order}
\usage{
fine_separate(meshcode = NULL, order = c("harf", "quarter"), ...)
}
\arguments{
\item{meshcode}{mesh code}

\item{order}{Choose mesh order}

\item{...}{other parameters for paste}
}
\value{
character vector
}
\description{
Return contains fine mesh codes
}
\examples{
fine_separate(52350400, "harf")
fine_separate(52350400, "quarter")
fine_separate(52350400, "quarter", collapse = ",")
}

