% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_rectangle.R
\name{mesh_rectangle}
\alias{mesh_rectangle}
\title{mesh rectangle check}
\usage{
mesh_rectangle(df, code = "mesh_code", view = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{code}{mesh code}

\item{view}{option.}
}
\description{
Multiple mesh coverd area.
}
\examples{
\dontrun{
library(dplyr)
mesh_rectangle(pref_mesh(33), code = "id", view = FALSE)
}
}

