% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_rectangle.R
\name{mesh_rectangle}
\alias{mesh_rectangle}
\title{mesh rectangle check}
\usage{
mesh_rectangle(df, code = "mesh_code", view = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{code}{mesh code}

\item{view}{option.}
}
\description{
Multiple mesh coverd area.
}
\examples{
\dontrun{
mesh_rectangle(data.frame(mesh = 6041,
lat_center = 40.3333333333,
long_center = 141.5,
lat_error =  0.33,
long_error = 0.5), "mesh", view = FALSE)
}
}
