% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_latlon.R
\name{mesh_to_coords}
\alias{mesh_to_coords}
\title{Get from mesh code to latitude and longitude}
\usage{
mesh_to_coords(meshcode, ...)
}
\arguments{
\item{meshcode}{\code{integer}. mesh code}

\item{...}{other parameters}
}
\description{
mesh centroid
}
\examples{
mesh_to_coords(64414277)
}
\references{
Akio Takenaka: \url{http://takenaka-akio.org/etc/j_map/index.html}
}
\seealso{
\code{\link[=coords_to_mesh]{coords_to_mesh()}} for convert from coordinates to meshcode
}
