% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{meshcode_set}
\alias{meshcode_set}
\title{Export 80km meshcode vectors}
\usage{
meshcode_set(mesh_size = c("80km", "10km", "1km"))
}
\arguments{
\item{mesh_size}{Export mesh size from 80km to 1km.}
}
\description{
Unique 176 meshcodes. 
The output code may contain values not found in the actual mesh code.
}
