% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{meshcode_set}
\alias{meshcode_set}
\title{Export meshcode vectors ranges 80km to 1km.}
\usage{
meshcode_set(mesh_size = c(80, 10, 5, 1))
}
\arguments{
\item{mesh_size}{Export mesh size from 80km to 1km.}
}
\description{
Unique 176 meshcodes.
The output code may contain values not found in the actual mesh code.
}
\examples{
meshcode_set(mesh_size = 80)
}
