% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbor_mesh.R
\name{neighbor_mesh}
\alias{neighbor_mesh}
\alias{find_neighbor_mesh}
\title{Find out neighborhood meshes collection}
\usage{
neighbor_mesh(meshcode, contains = TRUE)

find_neighbor_mesh(meshcode = NULL, contains = TRUE)
}
\arguments{
\item{meshcode}{\code{character}. mesh code}

\item{contains}{logical. contains input meshcode (default \code{TRUE})}
}
\value{
mesh code vectors (\code{character})
}
\description{
input should use meshcode under the 1km mesh size.
}
\examples{
neighbor_mesh(53394501)
neighbor_mesh(533945011)
neighbor_mesh(533945011, contains = FALSE)
}
