% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_mesh.R
\name{pref_mesh}
\alias{pref_mesh}
\title{Collect prefecture 1km mesh data}
\usage{
pref_mesh(code = NULL)
}
\arguments{
\item{code}{prefecture code (jiscode)}
}
\description{
Japanese prefectures administrations
}
\examples{
head(pref_mesh(33))
}

